requirements.txt
# Parsing
py-tree-sitter==0.20.4
tree-sitter-python==0.20.4

# Vector Search
chromadb==0.4.22
sentence-transformers==2.3.1

# Keyword Search
rank-bm25==0.2.2

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# LLM Client
requests==2.31.0

# CLI
click==8.1.0
rich==13.7.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-benchmark==4.0.0

# Profiling
memory-profiler==0.61.0
snakeviz==2.2.0

# Code Quality
ruff==0.1.13

# TUI Enhancement
textual==0.39.0
